+++ r4/src/test/java/io/github/jy95/fds/r4/translators/MaxDosePerLifetimeTest.java
@@ -1,52 +1,43 @@
 package io.github.jy95.fds.r4.translators;
 
-import io.github.jy95.fds.r4.DosageAPIR4;
-import io.github.jy95.fds.r4.AbstractFhirTest;
 import io.github.jy95.fds.common.types.DisplayOrder;
+import io.github.jy95.fds.common.types.DosageAPI;
+import io.github.jy95.fds.r4.DosageAPIR4;
+import io.github.jy95.fds.r4.config.FDSConfigR4;
+import io.github.jy95.fds.translators.AbstractMaxDosePerLifetimeTest;
 import org.hl7.fhir.r4.model.Dosage;
 import org.hl7.fhir.r4.model.Quantity;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.MethodSource;
 
+import java.util.List;
 import java.util.Locale;
-import java.util.concurrent.ExecutionException;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
+public class MaxDosePerLifetimeTest extends AbstractMaxDosePerLifetimeTest<FDSConfigR4, Dosage> {
 
-public class MaxDosePerLifetimeTest extends AbstractFhirTest {
-
-    @ParameterizedTest
-    @MethodSource("localeProvider")
-    void testNoMaxDosePerLifetime(Locale locale) throws ExecutionException, InterruptedException {
-        Dosage dosage = new Dosage();
-        DosageAPIR4 dosageUtils = getDosageAPI(locale, DisplayOrder.MAX_DOSE_PER_LIFETIME);
-        String result = dosageUtils.asHumanReadableText(dosage).get();
-        assertEquals("", result);
-    }
-
-    @ParameterizedTest
-    @MethodSource("localeProvider")
-    void testWithMaxDosePerLifetime(Locale locale) throws ExecutionException, InterruptedException {
+    @Override
+    protected Dosage generateWithMaxDosePerLifetime() {
         Dosage dosage = new Dosage();
         Quantity quantity = new Quantity();
         quantity.setValue(50);
         quantity.setUnit("mg");
         dosage.setMaxDosePerLifetime(quantity);
-        DosageAPIR4 dosageUtils = getDosageAPI(locale, DisplayOrder.MAX_DOSE_PER_LIFETIME);
-        String result = dosageUtils.asHumanReadableText(dosage).get();
-        String expected = getExpectedText(locale);
-        assertEquals(expected, result);
+        return dosage;
     }
 
-    private String getExpectedText(Locale locale) {
-        if (locale.equals(Locale.ENGLISH)) {
-            return "up to a maximum of 50 mg for the lifetime of the patient";
-        } else if (locale.equals(Locale.FRENCH)) {
-            return "jusqu’à un maximum de 50 mg pour la durée de vie du patient";
-        } else if (locale.equals(Locale.GERMAN)) {
-            return "bis zu einer maximalen Menge von 50 mg über die Lebenszeit des Patienten";
-        } else {
-            return "tot een maximum van 50 mg gedurende de levensduur van de patiënt";
-        }
+    @Override
+    public DosageAPI<FDSConfigR4, Dosage> getDosageAPI(Locale locale, DisplayOrder displayOrder) {
+        return new DosageAPIR4(FDSConfigR4.builder()
+                .displayOrder(List.of(displayOrder))
+                .locale(locale)
+                .build());
+    }
+
+    @Override
+    public DosageAPI<FDSConfigR4, Dosage> getDosageAPI(FDSConfigR4 config) {
+        return new DosageAPIR4(config);
+    }
+
+    @Override
+    public Dosage generateEmptyDosage() {
+        return new Dosage();
     }
 }
